+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Big convenience enhancements to DOM interface. An improved pull parser interface
 - **Improved:** Various documentation improvements.
 - **Change:** Minimum DMDFE bumped to 2.069.
 - **Change:** Cleaned up the names of the [exceptions](http://semitwist.com/sdlang-d/sdlang/exception.html).
-- **Change:** Deprecated `Tag.fullName`. Use `Tag.getFullName().toString()` instead.
+- **Change:** Deprecated `Tag.fullName` and `Attribute.fullName`. Use `Tag.getFullName().toString()` and `Attribute.getFullName().toString()` instead.
 - **Change:** [#17](https://github.com/Abscissa/SDLang-D/issues/17): Remove unnecessary FileStart and FileEnd events.
 - **Change:** Pull parser: [ParserEvent](http://semitwist.com/sdlang-d/sdlang/parser/ParserEvent.html) is now a [TaggedAlgebraic](https://github.com/s-ludwig/taggedalgebraic), instead of a [std.variant.Algebraic](http://dlang.org/phobos/std_variant.html#.Algebraic). This enables use of `final switch` and provides simpler syntax.
 - **Fixed:** Fixed building the CLI tool via dub.